Extract window frame reconcile helpers
